UPD: Spliting Base object, by sending it to base-model_file and changing docker-compose file to run uviconr from there in order to run in debug mode
