+++ Back_End/flask-server/models/__init__.py
@@ -3,10 +3,6 @@
 initialize the models package
 """
 
-from sqlalchemy.ext.declarative import declarative_base
-
-# Declare Base for SQL Alchemy table creation
-Base = declarative_base()
 from models.engine.db import DBStorage
 
 # creates instance for DBStorage that allows communication with DB through this instance
@@ -14,4 +10,4 @@ storage = DBStorage()
 
 # Starts an SQL session and creates the tables in DB
 storage.reload()
-print("hice reload")
+print("Reloading DB Storage")
