+++ Back_End/flask-server/models/base_model.py
@@ -1,0 +1,6 @@
+#!/usr/bin/python3
+
+from sqlalchemy.ext.declarative import declarative_base
+# Declare Base for SQL Alchemy table creation
+
+Base = declarative_base()
