+++ Back_End/flask-server/models/budget.py
@@ -5,7 +5,7 @@
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, DECIMAL
 from datetime import datetime
-from models import Base
+from models.base_model import Base
 # Base = declarative_base()
 
 
@@ -15,7 +15,6 @@ class Budget(Base):
     __tablename__ = 'budgets'
     budget_id = Column(Integer, primary_key=True, autoincrement=True)
     date = Column(DateTime, default=datetime.now, nullable=False)
-    value = Column(DECIMAL(15,2), nullable=False, default=0.00)
+    value = Column(DECIMAL(15, 2), nullable=False, default=0.00)
     category_id = Column(Integer, ForeignKey(
         'categories.category_id'), nullable=False)
-
