+++ Back_End/flask-server/models/user.py
@@ -5,7 +5,7 @@
 from sqlalchemy import Column, Integer, String
 from sqlalchemy.orm import relationship
 from sqlalchemy.ext.declarative import declarative_base
-from models import Base
+from models.base_model import Base
 
 # Base = declarative_base()
