+++ Back_End/flask-server/saveapp.py
@@ -2,6 +2,3 @@ from app.app import *
 
 if __name__ == '__main__':
     """ Main Function """
-    host = '0.0.0.0'
-    port = '5000'
-    uvicorn.run("app.app:app", host=host, port=port)
